[CL-2110] Move verifications status
